requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
scikit-learn==1.5.2
pandas==2.2.3
joblib==1.4.2
gunicorn==21.2.0
fpdf2
icalendar
